Serialize offer creation

+++ vexl/Repositories/OfferRepository.swift
@@ -13,7 +13,7 @@ import UIKit
 protocol OfferRepositoryType {
     func createOffer(keys: ECCKeys, locations: [OfferLocation], provider: @escaping (ManagedOffer) -> Void) -> AnyPublisher<ManagedOffer, Error>
 
-    func update(offer: ManagedOffer, locations: [OfferLocation], provider: @escaping (ManagedOffer) -> Void) -> AnyPublisher<ManagedOffer, Error>
+    func update(offer: ManagedOffer, locations: [OfferLocation]?, provider: @escaping (ManagedOffer) -> Void) -> AnyPublisher<ManagedOffer, Error>
 
     func createOrUpdateOffer(offerPayloads: [OfferPayload]) -> AnyPublisher<[ManagedOffer], Error>
 
@@ -63,39 +63,33 @@ class OfferRepository: OfferRepositoryType {
             offer.user = userRepository.getUser(for: context)
             offer.locations = NSSet(array: managedLocations)
 
-            let item = ManagedSyncItem(context: context)
-            item.type = .insert
-            item.offer = offer
-
             return offer
         }
     }
 
     func update(
         offer: ManagedOffer,
-        locations: [OfferLocation],
+        locations: [OfferLocation]?,
         provider: @escaping (ManagedOffer) -> Void
     ) -> AnyPublisher<ManagedOffer, Error> {
         persistence.update(context: persistence.viewContext) { context in
-            let managedLocations = locations.map { location -> ManagedOfferLocation in
-                let managedLocation = ManagedOfferLocation(context: context)
-                managedLocation.lat = location.latitude
-                managedLocation.lon = location.longitude
-                managedLocation.city = location.city
-                return managedLocation
-            }
+            if let locations = locations {
+                let managedLocations = locations.map { location -> ManagedOfferLocation in
+                    let managedLocation = ManagedOfferLocation(context: context)
+                    managedLocation.lat = location.latitude
+                    managedLocation.lon = location.longitude
+                    managedLocation.city = location.city
+                    return managedLocation
+                }
 
-            (offer.locations as? Set<ManagedOfferLocation>)?.forEach { location in
-                context.delete(location)
-            }
+                (offer.locations as? Set<ManagedOfferLocation>)?.forEach { location in
+                    context.delete(location)
+                }
 
-            offer.locations = NSSet(array: managedLocations)
+                offer.locations = NSSet(array: managedLocations)
+            }
 
             provider(offer)
-
-            let item = ManagedSyncItem(context: context)
-            item.type = .update
-            item.offer = offer
 
             return offer
         }
